refactor: 당첨 번호 기능 로또 기능으로 리팩토링
