+++ src/main/java/lotto/WinNumber.java
@@ -10,7 +10,7 @@ import java.util.stream.Collectors;
 
 public class WinNumber {
 
-    private final List<Integer> numbers;
+    private final Lotto numbers;
 
     public void requestWinLottoNumber() {
         System.out.println("당첨 번호를 입력해 주세요.");
@@ -18,10 +18,7 @@ public class WinNumber {
 
     public WinNumber(String numbers) {
         List<Integer> tmpNumbers = validateNumberType(numbers);
-        validateNumberLength(tmpNumbers);
-        validateNumberReplicate(tmpNumbers);
-        validateNumberRange(tmpNumbers);
-        this.numbers = tmpNumbers;
+        this.numbers = new Lotto(tmpNumbers);
     }
 
     public List<Integer> validateNumberType(String numbers) {
@@ -32,31 +29,6 @@ public class WinNumber {
         catch (Exception e) {
             System.out.println("[ERROR] 로또 당첨 번호는 숫자여야 합니다.");
             throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateNumberLength(List<Integer> numbers) {
-        if (numbers.size() != 6) {
-            System.out.println("[ERROR] 로또 당첨 번호는 6개여야 합니다.");
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateNumberReplicate(List<Integer> numbers) {
-        Set<Integer> independentNumbers = new HashSet<>(numbers);
-
-        if (independentNumbers.size() != numbers.size()) {
-            System.out.println("[ERROR] 로또 번호는 중복될 수 없습니다.");
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateNumberRange(List<Integer> numbers) {
-        for (int number : numbers) {
-            if (number < 1 || number > 45) {
-                System.out.println("[ERROR] 로또 번호는 1~45 사이의 숫자여야 합니다.");
-                throw new IllegalArgumentException();
-            }
         }
     }
 }
